Added generic geometry generation class.
First part of the dynamic geometry generation (the one that does Marching Cubes).

Started adding parts from RenderPassGeometryGeneration. WIP

Further copying. WIP

For now, the _computeCallNormalRef and _computeCallPerlinRef don't work now. WIP.

Now, initialization of compute calls works!

Changed the indents.

Adding the structure where we can add more shaders.

Added DynamicGeneratedMesh struct with a constructor.

Added addingGeneratedMesh with r-value reference to the DynamicGeneratedMesh class.

Changed std::string& to Name&.

Added a loop which goes through shaders.

Added a loop through all of the shaders for init() and postInit().

Added forward declaration and arrays of BuffersRefs, ImageRefs, PipelineRefs, ComputeCallRefs.

Started adding the buffers as members of DynamicGeneratedMeshes struct.

Changing every buffer to be a part of DynamicGeometryGeneration structure. Does not work yet but it compiles.

Further adding every buffer to the DynamicGeometryGeneration.

Further moving the refs to the DynamicGeometryGeneration structure. For now it works.

Further moving the references into the DynamicGeometryGeneration.

Now all of the references to the buffers, images, pipelines and compute calls are inside the DynamicGeometryGeneration structure.

Further refactor, still works.

Refactored the indents of the for loop.

Refactored the indents in the for loop for init() but it still is not good.

It works!

Cleaned up a little bit.

Cleaned-up the shader names and added the mesh name.

Now rendering two shapes as separate meshes! So it works!

Cleaned up the uneeded shaders.

Added julia to the collection.

Almost done - need to use separate gbuffer shader pass only for pbr materials.

Added a new separated material pass for the fractals. It works and is beautiful.

Now motion of the julia fractal works.

Cleaning up.

Added static meshes, whhich do not change during the execution of the program. For static meshes the binormal and tangents don't work (checked that and I am sure).

Static mesh started to work!

Further improvement. Now only 2 passes are needed to generate a proper mesh.

Refactored *perlin* -> *SDFGeneration*.

Cleaning up.

A little refactor to update the memory for each dynamic mesh and avoiding it for the static ones.